move tests from test_regression.py to appropriate files
